apps/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
certifi==2025.4.26
click==8.1.8
coverage==7.8.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
packaging==25.0
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dotenv==1.0.1
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.3
ruff==0.11.7
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.37.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
